+++ src/oimo/dynamics/TimeStep.java
@@ -7,18 +7,18 @@ public class TimeStep {
 	/**
 	 * The time step of simulation.
 	 */
-	public float dt;
+	public double dt;
 
 	/**
 	 * The inverse time step of simulation, equivalent to simulation FPS.
 	 */
-	public float invDt;
+	public double invDt;
 
 	/**
 	 * The ratio of time steps. Defined by current time step divided by previous
 	 * time step.
 	 */
-	public float dtRatio;
+	public double dtRatio;
 
 	public TimeStep() {
 		dt = 0;
