+++ src/oimo/dynamics/callback/RayCastClosest.java
@@ -27,7 +27,7 @@ public  class RayCastClosest implements RayCastCallback {
 	/**
 	 * The ratio of the position the ray hit from the start point to the end point.
 	 */
-	public float fraction;
+	public double fraction;
 
 	/**
 	 * Whether the ray hit any shape in the world.
